Splits code related to different threads into different files.
